+++ experiments/browser_based_querying/src/hackernews/fetcher.ts
@@ -1,4 +1,4 @@
-import { SendableSyncContext, SyncContext } from './sync';
+import { SendableSyncContext, SyncContext } from '../sync';
 
 interface ChannelData {
   input: string;
