+++ experiments/browser_based_querying/src/hackernews/schema.graphql
